feat(local-dev): Add a fixture with some init data
When setting the api for local developement(or for just getting the feel of it)
We need some initial data like admin user, a project a feature etc.
This commit tries to address that issue by automating all that boring stuff
with the help of the fixtures.
To load the data on fresh database run the below command.
python manage.py loaddata  app/fixtures/inti_data.yaml

Admin user created by this fixture have the following credentials:
email: local@flagsmith.com
password: test123