improvement/sentry cors (#475)
* Update common.py (#449)

The way Sentry works in the browser is that it injects itself into the XMLHttpRequest function and adds a sentry-trace header there. This ends up causing calls to the Flagsmith api to fail. I think the proper way to do it is just allow the sentry flags to pass through

Optionally just allow ALL headers. I am not sure why there are header restrictions in the first place.

* Black formatting

Co-authored-by: Andrew Nagy <tm1000@users.noreply.github.com>