fix(audit/signal/task): fix race condition
The environment document in dynamodb is updated based on the post_save signal from the audit log
When creating a new feature, the feature states are created after the feature has been created.
i.e: the below task gets created/scheduled before feature states are created
Usually, there is enough time for the main thread to create the feature states
before the task is executed, but not always.
In those cases, we send the environment document to dynamodb without any feature states for the new feature.
In order to avoid this, either we need to update environment
document when creating feature states
or delay the execution of this task
We prefer to delay the execution of the task because of it's low surface area